3/12 Update: Revised main.c, made can_manager.c/h
main.c - Added global variables from the old version and tried to comment out ones that seemed obsolete or not needed(needs revision). Also tried to rewrite can_send_temp() for a dynamic number of subpacks, and changed code for time so it measured the time of cycle(used in Kalman Filter for SOC est.)

can_manager.c - Copied from the old version of the code on the FRUCD github repos and needed for the can functions in main.c. Deleted functions relating to current and attempted to rewrite the can_send_temp() function.

can_manager.h - Also copied from the old code, made changes according to the changes in can_manager.c and deleted an enum that was in conflict from cell_interface.h that needed to be included for the BAT_SUBPACK_t struct to be defined in the can_send_temp function.

Also, this version has around 10 errors in LTC6811.h, mostly about "redefining" or "redeclaring" certain things.